ecto_factory v0.0.1 EctoFactory

Easily generate structs based on your ecto schemas.

View Hex docs

Installation

Add ecto_factory to your list of dependencies in mix.exs:

def deps do
  [{:ecto_factory, "~> 0.0.1"}]
end

Usage

Using the following MyApp.User module:

defmodule MyApp.User do
  use Ecto.Schema

  schema "users" do
    field :age, :integer
    field :username, :string
    field :date_of_birth, Ecto.DateTime
  end
end

Configure ecto_factory in a few different ways.

#./config.exs

config :ecto_factory, factories: [
  :user, MyApp.User 

  :user_with_defaults, { MyApp.User, [
    age: 99,
    username: "mrmicahcooper"
    date_of_birth: Ecto.DateTime.cast!("2012-12-12"),
  ] }
]

And that’s it. Now use EctoFactory.build to create structs.

EctoFactory.build(:user) 
#=> %MyApp.User{age: 1, username: "username, date_of_birth: #Ecto.DateTime<2016-06-14T17:03:22Z>}

EctoFactory.build(:user, username: "hashrocket")
#=> %MyApp.User{age: 1, username: "hashrocket, date_of_birth: #Ecto.DateTime<2016-06-14T17:03:22Z>}

Ecto.Factory.build(:user_with_defaults)
#=> %MyApp.User{age: 99, username: "mrmicahcooper, date_of_birth: #Ecto.DateTime<2012-12-12T00:00:00Z>}

EctoFactory uses the fields you’ve defined in your schema to create some basic data that can be easily overwritten with a very small amount of configuration.

Functions

build(model_name, attrs \\ %{})

Create a struct of the passed in factory

After configuring a factory

config :ecto_factory, factories: [
  user_with_default_username: { User,
    username: "mrmicahcooper"
  }
]

You can build a struct with the attributes from my factory as defaults.

iex> EctoFactory.build(:user_with_default_username)
%User{
  age: 1,
  username: "mrmicahcooper",
  date_of_birth: Ecto.DateTime.utc
}

And you can pass in your own attributes of course:

iex> EctoFactory.build(:user, age: 99, username: "hashrocket")
%User{
  age: 99,
  username: "hashrocket",
  date_of_birth: Ecto.DateTime.utc
}